requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
openai>=1.0.0
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0